Have publish.sh handle failure better.
Instead of pushing an empty branch to the site, abort.

+++ scripts/publish.sh
@@ -1,5 +1,11 @@
 #!/bin/sh
 
+
+if ! (cat config.toml > /dev/null); then
+    echo "Must run from repo root"
+    exit 1
+fi
+
 if [[ $(git status -s) ]]
 then
     echo "The working directory is dirty. Please commit any pending changes."
@@ -13,13 +19,19 @@ git worktree prune
 rm -rf .git/worktrees/public/
 
 echo "Checking out gh-pages branch into public"
-git worktree add -B gh-pages public origin/gh-pages
+if ! git worktree add -B gh-pages public origin/gh-pages ; then
+    echo "Failed to create worktree."
+    exit 1
+fi
 
 echo "Removing existing files"
 rm -rf public/*
 
 echo "Generating site"
-hugo
+if ! hugo ; then
+    echo "Build failed"
+    exit 1
+fi
 
 echo "Updating gh-pages branch"
 cd public && git add --all && git commit -m "Publishing to gh-pages (publish.sh)"
